Merged PR 38: Add simple and lightweight logging functionality
Constructor options now take an additional and optional parameter called logLevel. The logLevel parameter is interpreted by _setLoging() that, where required, replaces console.log and/or console.error with an empty function. Effectively disabling synchronous and blocking writes to stdOut and stdError.

Code has been liberally sprinkled with informative console.log and console.error commands.

Also, included an earlier commitment of code that detects whether the function is executing offline (via serverless offline) or online (aka inside the real AWS Lambda environment). This information is used to determine whether to create the build and dist folders under `/tmp` (absolute from root of the filesystem) or `../tmp` relative to code. Remember that in real Lambda the only part of the filesystem we are granted write access to is `/tmp`.

Related work items: #315