Merged PR 47: Fix for bug #351 and sets up the code for future S3 source of stack
The bug itself was due to the difference between local (offline) and AWS (online) environments. When online AWS only allows writing to `/tmp`. When offline I wanted to make it easier to see and work with files so instead of writing to our workstation `/tmp` it was supposed to write to a dist folder relative to the project. Along the way the logic for checking environments, and setting all the various path occurrences in code got messed up. That is now working for local.

I used this as an opportunity to start setting up the online world (AWS) to pull the stack files from an S3 bucket. To reiterate, on local dev they currently pull from a relative stack directory. Once we get past this commit and a corresponding commit in YMWA to ensure it is up to date we should start testing by manually copying files from our local relative stack directory to the S3 bucket and manually invoke the Lambda on AWS. Once we know that is working we can move the stack files to a new www.yourmentalwealthadvisors.com repo as they will not need to be deployed using `sls deploy` anymore.

Testing Slack webhook. Ignore this line.

Related work items: #351