Merged PR 85: Fixes issue with Amplify being current -1
There was code leftover from a previous incarnation when things were less streams-based. It was benign on local dev because of the amount of time it takes to fetch assets across the internet from S3. In the AWS environment, however, assets are retrieved much faster and a race condition was hampered by the code. The code has now been removed.

Also cleaned up some formatting in Vinyl-S3.js

Related work items: #418